Use different sources for reader and writer
